assignment_4/requirements.txt
facenet-pytorch
matplotlib
numpy
pandas
Pillow
tensorflow
tensorflow_hub
tf-keras
torch
tqdm
transformers




